+++ crates/objc2/src/runtime.rs
@@ -197,6 +197,14 @@ impl Sel {
         let name = unsafe { CStr::from_ptr(ptr) };
         str::from_utf8(name.to_bytes()).unwrap()
     }
+
+    pub(crate) fn number_of_arguments(self) -> usize {
+        self.name()
+            .as_bytes()
+            .iter()
+            .filter(|&&b| b == b':')
+            .count()
+    }
 }
 
 // `ffi::sel_isEqual` is just pointer comparison on Apple (the documentation
@@ -983,6 +991,7 @@ mod tests {
 
     use super::*;
     use crate::test_utils;
+    use crate::MessageReceiver;
     use crate::{msg_send, sel};
 
     #[test]
@@ -999,6 +1008,14 @@ mod tests {
         test_sel!("abc:", abc:);
         test_sel!("abc:def:", abc:def:);
         test_sel!("abc:def:ghi:", abc:def:ghi:);
+        test_sel!("functionWithControlPoints::::", functionWithControlPoints::::);
+        test_sel!("initWithControlPoints::::", initWithControlPoints::::);
+        test_sel!("setFloatValue::", setFloatValue::);
+        test_sel!("isSupported::", isSupported::);
+        test_sel!("addEventListener:::", addEventListener:::);
+        test_sel!("test::arg::", test::arg::);
+        test_sel!("test::::with::spaces::", test : :: : with : : spaces : :);
+        test_sel!("a::b:", a::b:);
     }
 
     #[test]
@@ -1007,6 +1024,8 @@ mod tests {
         assert_eq!(sel.name(), "");
         let sel = Sel::register(":");
         assert_eq!(sel.name(), ":");
+        let sel = Sel::register("::");
+        assert_eq!(sel.name(), "::");
     }
 
     #[test]
@@ -1071,6 +1090,7 @@ mod tests {
 
         assert!(cls.responds_to(sel!(foo)));
         assert!(cls.responds_to(sel!(setBar:)));
+        assert!(cls.responds_to(sel!(test::test::)));
         assert!(!cls.responds_to(sel!(abc)));
         assert!(!cls.responds_to(sel!(addNumber:toNumber:)));
 
@@ -1080,6 +1100,7 @@ mod tests {
         // The metaclass of a root class is a subclass of the root class
         assert_eq!(metaclass.superclass().unwrap(), cls);
         assert!(metaclass.responds_to(sel!(addNumber:toNumber:)));
+        assert!(metaclass.responds_to(sel!(test::test::)));
         // TODO: This is unexpected!
         assert!(metaclass.responds_to(sel!(foo)));
 
@@ -1222,5 +1243,20 @@ mod tests {
             format!("{:?}", &*object),
             format!("<CustomObject: {:p}>", &*object)
         );
+    }
+
+    #[test]
+    fn test_multiple_colon() {
+        let class = test_utils::custom_class();
+        let res: i32 = unsafe {
+            MessageReceiver::send_message(class, sel!(test::test::), (1i32, 2i32, 3i32, 4i32))
+        };
+        assert_eq!(res, 10);
+
+        let obj = test_utils::custom_object();
+        let res: i32 = unsafe {
+            MessageReceiver::send_message(&obj, sel!(test::test::), (1i32, 2i32, 3i32, 4i32))
+        };
+        assert_eq!(res, 24);
     }
 }