+++ crates/objc2-proc-macros/CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased - YYYY-MM-DD
 
+### Fixed
+* Allow all types of tokens in internal macro.
+
 
 ## 0.1.0 - 2022-07-19